refactor(validation): extract schema loading
